feat: 회원탈퇴 다이얼로그

+++ app/src/main/java/com/example/sharedtaxitogether/ProfileFragment.kt
@@ -1,12 +1,14 @@
 package com.example.sharedtaxitogether
 
 import android.content.Context
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.room.Room
 import com.example.sharedtaxitogether.databinding.FragmentProfileBinding
@@ -26,25 +28,48 @@ class ProfileFragment : Fragment() {
     }
 
     private fun bind() {
-        binding.logoutTextView.setOnClickListener {
-            val user = room.userDao().getUser()
-            Log.d("profileFragment", user.toString())
-            room.userDao().delete(user)
+        binding.modifyInfoTextView.setOnClickListener {
 
-            startActivity(Intent(mainActivity,LoginActivity::class.java))
+        }
+        binding.logoutTextView.setOnClickListener {
+            logout()
         }
         binding.withdrawTextView.setOnClickListener {
-            val token = room.userDao().getUid()
-            db.collection("users").document(token)
-                .delete()
-                .addOnSuccessListener {
-                    Log.d("profileFragment","회원탈퇴 성공")
-                    startActivity(Intent(mainActivity,LoginActivity::class.java))
-                }
+            withdraw()
         }
         binding.profileImgView.setOnClickListener {
             // TODO 프로필 변경
+
         }
+    }
+
+    private fun withdraw() {
+        val builder = AlertDialog.Builder(mainActivity)
+        builder.setTitle("회원탈퇴")
+            .setMessage("${room.userDao().getNickname()}님 정말 탈퇴하시겠습니까?")
+            .setPositiveButton("탈퇴하기",
+                DialogInterface.OnClickListener { _, _ ->
+                    val token = room.userDao().getUid()
+                    db.collection("users").document(token)
+                        .delete()
+                        .addOnSuccessListener {
+                            Log.d(TAG, "회원탈퇴 성공")
+                            startActivity(Intent(mainActivity, LoginActivity::class.java))
+                        }
+                })
+            .setNegativeButton("취소",
+                DialogInterface.OnClickListener { _, _ ->
+                    Log.d(TAG, "회원탈퇴 취소")
+                })
+        builder.show()
+    }
+
+    private fun logout() {
+        val user = room.userDao().getUser()
+        Log.d(TAG, user.toString())
+        room.userDao().delete(user)
+
+        startActivity(Intent(mainActivity, LoginActivity::class.java))
     }
 
     override fun onCreateView(
@@ -80,4 +105,7 @@ class ProfileFragment : Fragment() {
         return binding.root
     }
 
+    companion object {
+        private const val TAG = "profileFragment"
+    }
 }
